changes to add roles in add and update command handlers

+++ Src/csharp-mono/Application.Courses/Commands/DeleteCourseDefination/DeleteCourseDefinationCommandHandler.cs
@@ -1,6 +1,8 @@
 using System;
 using System.Threading;
 using System.Threading.Tasks;
+using Application.Helpers;
+using Application.Interfaces;
 using Bridge.Application.Interfaces;
 using Bridge.Application.Models;
 using Bridge.Domain.Entities;
@@ -12,38 +14,55 @@ namespace Application.Courses.Commands.DeleteCourseDefination
     public class DeleteCourseDefinationCommandHandler : IRequestHandler<DeleteCourseDefinationCommand, ApiResponse>
     {
         private readonly IBridgeDbContext _dbContext;
-        public DeleteCourseDefinationCommandHandler(IBridgeDbContext dbContext)
+        private readonly ICurrentUserService _userService;
+        private readonly IUserHelper _userHelper;
+
+
+        public DeleteCourseDefinationCommandHandler(IBridgeDbContext dbContext, ICurrentUserService userService,IUserHelper userHelper)
         {
             _dbContext = dbContext;
+            _userService = userService;
+            _userHelper = userHelper;
         }
         public async Task<ApiResponse> Handle(DeleteCourseDefinationCommand request, CancellationToken cancellationToken)
         {
             ApiResponse res = new ApiResponse();
             try
             {
-                CourseDefination ifExistCourseDefinition = await _dbContext.CourseDefination.FirstOrDefaultAsync(x => x.Id == request.Id );
-                if (ifExistCourseDefinition != null)
+                if (_userService.RoleList.Contains(Roles.admin.ToString()))
                 {
-                    // ifExistCourseDefinition.IsDeleted = true;
-                    ifExistCourseDefinition.LastModifiedBy = 0;
-                    ifExistCourseDefinition.LastModifiedOn = DateTime.UtcNow;
-                    await _dbContext.SaveChangesAsync(cancellationToken);
-
-                    DeleteCourseDefinationVm vmObj = new DeleteCourseDefinationVm()
+                    var userId = await _userHelper.getUserId(_userService.UserId.ToString());
+                  
+                    CourseDefination ifExistCourseDefinition = await _dbContext.CourseDefination.FirstOrDefaultAsync(x => x.Id == request.Id);
+                    if (ifExistCourseDefinition != null)
                     {
-                        Id = ifExistCourseDefinition.Id,
-                        GradeId = ifExistCourseDefinition.Id,
-                        CourseId = ifExistCourseDefinition.CourseId,
-                        Subject = ifExistCourseDefinition.Subject,
-                        BasePrice = ifExistCourseDefinition.BasePrice,
-                    };
-                    res.data = vmObj;
-                    res.response_code = 0;
-                    res.message = "CourseDefination Deleted";
-                    res.status = "Success";
-                    res.ReturnCode = 200;
+                        // ifExistCourseDefinition.IsDeleted = true;
+                        ifExistCourseDefinition.LastModifierUserId = int.Parse(userId);
+                        ifExistCourseDefinition.LastModificationTime = DateTime.UtcNow.ToString();
+                        await _dbContext.SaveChangesAsync(cancellationToken);
+
+                        DeleteCourseDefinationVm vmObj = new DeleteCourseDefinationVm()
+                        {
+                            Id = ifExistCourseDefinition.Id,
+                            GradeId = ifExistCourseDefinition.Id,
+                            CourseId = ifExistCourseDefinition.CourseId,
+                            Subject = ifExistCourseDefinition.Subject,
+                            BasePrice = ifExistCourseDefinition.BasePrice,
+                        };
+                        res.data = vmObj;
+                        res.response_code = 0;
+                        res.message = "CourseDefination Deleted";
+                        res.status = "Success";
+                        res.ReturnCode = 200;
+                    }
                 }
-                
+                else
+                {
+                    res.response_code = 1;
+                    res.message = "You are not authorized.";
+                    res.status = "Unsuccess";
+                    res.ReturnCode = 401;
+                }
             }
             catch (Exception ex)
             {
